rename kafka2 pkg to kafka.internal
* will use as base for implementation
