+++ core/src/main/scala/com/softwaremill/react/kafka/internal/ConsumerProvider.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2014 - 2016 Softwaremill <http://softwaremill.com>
  * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package com.softwaremill.react.kafka2
+package com.softwaremill.react.kafka.internal
 
 import org.apache.kafka.clients.consumer.KafkaConsumer
 import org.apache.kafka.clients.producer.ProducerConfig
