+++ core/src/test/scala/com/softwaremill/react/kafka/internal/ConsumerTest.scala
@@ -1,4 +1,8 @@
-package com.softwaremill.react.kafka2
+/*
+ * Copyright (C) 2014 - 2016 Softwaremill <http://softwaremill.com>
+ * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
+ */
+package com.softwaremill.react.kafka.internal
 
 import java.util.{Map => JMap}
