+++ core/src/test/scala/com/softwaremill/react/kafka/internal/ProducerTest.scala
@@ -2,7 +2,7 @@
  * Copyright (C) 2014 - 2016 Softwaremill <http://softwaremill.com>
  * Copyright (C) 2016 Lightbend Inc. <http://www.lightbend.com>
  */
-package com.softwaremill.react.kafka2
+package com.softwaremill.react.kafka.internal
 
 import java.util.concurrent.{CompletableFuture, TimeUnit}
